backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
weasyprint>=62.3
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
python-dotenv>=1.0.1
jinja2>=3.1.4
pydantic>=2.10.0
python-multipart>=0.0.6

